// Edukop-app/src/app/dynamic-form/model/formType.ts
export enum formType {
    TEXT = 'Text',
    TEXTAREA = 'Textarea',
    NUMBER = 'Number',
    EMAIL = 'Email',
    DROPDOWN = 'Dropdown',
    OPTIONS = 'Options',
    CHECKBOXES = 'CheckBoxes',
    BOOLEAN = 'Boolean',
    RANGE = 'Range',
    DOCUMENT_UPLOAD = 'Document upload',
    LOGO = 'Logo',
    WEBSITE = 'Website',
    DATE = 'Date'
}